% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExploracionABA.R
\name{GetAllAreasSimplified}
\alias{GetAllAreasSimplified}
\title{Obtiene todas las areas disponibles}
\usage{
GetAllAreasSimplified()
}
\value{
Data table con todas las areas disponibles en la base de datos
}
\description{
Obtiene todas las areas disponibles
}
\examples{
all.areas <- GetAllAreasSimplified()
}
