% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExploracionABA.R
\name{GetAreasGenes}
\alias{GetAreasGenes}
\title{Obtiene los genes que se expresan en una area}
\usage{
GetAreasGenes(area.selected = "accumbens",
  selected.dataset = "dataset_adult")
}
\arguments{
\item{area.selected}{Nombre completo o nombre parcial del area de la que se quieren obtener los genes expresados}

\item{selected.dataset}{Selección de una de las tres bases de datos disponibles en ABA}
}
\value{
Cata.table con los genes expresados en el área seleccionada
}
\description{
Obtiene los genes que se expresan en una area
}
\examples{
GetAreasGenes(area.selected = 'accumbens', selected.dataset = "dataset_adult")
}
\seealso{
\code{\link{GetGenAreas}} para obtener areas
}
