% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExploracionABA.R
\name{GetDatasetAreasSimplified}
\alias{GetDatasetAreasSimplified}
\title{Obtiene todas las areas disponibles}
\usage{
GetDatasetAreasSimplified(selected.dataset = "dataset_adult")
}
\arguments{
\item{selected.dataset}{Selección de una de las tres bases de datos existente, esta función tambien acepta un data.frame
con los números de area que querramos identificar en una columna llamada structure.}
}
\value{
Data table con todas las areas disponibles en la base de datos
}
\description{
Obtiene todas las areas disponibles
}
\examples{
all.areas <- GetDatasetAreasSimplified()
}
