% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExploracionABA.R
\name{GetGenAreas}
\alias{GetGenAreas}
\title{Obtiene las areas relacionadas con un Gen específico}
\usage{
GetGenAreas(gene.name.pattern = NULL, ensembl.id = NULL,
  selected.dataset = "dataset_adult")
}
\arguments{
\item{gene.name.pattern}{Nombre completo o nombre parcial del gen en cuestión}

\item{ensembl.id}{Nombre completo o nombre parcial del identificador}

\item{selected.dataset}{Selección de una de las tres bases de datos disponibles en ABA}
}
\value{
Data.table con las areas en las que se expresa el gen seleccionado.
}
\description{
Con esta función se pueden buscar los genes estudiados en uno de los dataset, buscandolo por el nombre o por el
identificador ensembl.
}
\examples{
GetGenAreas(gene.name.pattern = 'A1BG',ensembl.id = 'ENSG', selected.dataset = "dataset_5_stages")
}
\seealso{
\code{\link{GetAreasGenes}} para obtener areas
}
