% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExploracionABA.R
\name{GetAreasGenes}
\alias{GetAreasGenes}
\title{Obtiene los genes que se expresan en una area}
\usage{
GetAreasGenes(structure.selected = NULL, structure.id = NULL,
  selected.dataset = "dataset_adult")
}
\arguments{
\item{structure.id}{Identificador del area de la que se quieren obtener los genes expresados}

\item{selected.dataset}{Selección de una de las tres bases de datos disponibles en ABA}

\item{area.selected}{Nombre completo o nombre parcial del area de la que se quieren obtener los genes expresados}
}
\value{
data.table con los genes expresados en el área seleccionada
}
\description{
A partir del area seleccionada se pueden identificar los genes expresados dentro de la base seleccionada.
}
\examples{
GetAreasGenes(structure.selected = 'accumbens', selected.dataset = "dataset_adult")
GetAreasGenes(structure.id = 4679, selected.dataset = "dataset_adult")
GetAreasGenes(structure.selected = 'accumbens', structure.id = 4679, selected.dataset = "dataset_adult")
}
\seealso{
\code{\link{GetGenAreas}} para obtener areas
}
