% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExploracionABA.R
\name{GetGenAreas}
\alias{GetGenAreas}
\title{Obtiene las areas relacionadas con un Gen específico}
\usage{
GetGenAreas(gene.name.pattern, selected.dataset = "dataset_adult")
}
\arguments{
\item{gene.name.pattern}{Nombre completo o nombre parcial del gen en cuestión}

\item{selected.dataset}{Selección de una de las tres bases de datos disponibles en ABA}
}
\value{
Data.table con las areas en las que se expresa el gen seleccionado.
}
\description{
Obtiene las areas relacionadas con un Gen específico
}
\examples{
GetGenAreas(gene.name.pattern = 'A1BG', selected.dataset = "dataset_5_stages")
}
